requirement.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-pagedown==0.1.1
django-ckeditor==5.2.1
markdown2==2.3.2
Pillow==4.0.0
Unidecode==0.4.20
django-blog-zinnia==0.18.1